# be-nest/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password/reset/request": {
      "post": {
        "operationId": "AuthController_requestReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password/reset/confirm": {
      "post": {
        "operationId": "AuthController_confirmReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/Bachelor/Add": {
      "post": {
        "operationId": "BachelorController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
    "/api/Bachelor/Update": {
      "put": {
        "operationId": "BachelorController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BachelorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
    "/api/Bachelor/UpdateListBachelor/{hallId}/{sessionId}": {
      "put": {
        "operationId": "BachelorController_updateMany",
        "parameters": [
          {
            "name": "hallId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
    "/api/Bachelor/Delete/{studentCode}": {
      "delete": {
        "operationId": "BachelorController_delete",
        "parameters": [
          {
            "name": "studentCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
...
    "/api/Bachelor/GetByHallSession/{hallId}/{sessionId}": {
      "get": {
        "operationId": "BachelorController_byHallSession",
        "parameters": [
          {
            "name": "hallId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
    "/api/Bachelor/UpdateBachelorToTempSession/{studentCode}": {
      "put": {
        "operationId": "BachelorController_moveToTemporary",
        "parameters": [
          {
            "name": "studentCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToTemporarySessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
    "/api/Bachelor/TransferLateStudent": {
      "put": {
        "operationId": "BachelorController_transferLateStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferLateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bachelor"
        ]
      }
    },
...
  },
  "info": {
    "title": "Convocation Day API",
    "description": "NestJS compatibility API for the Convocation Day frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmPasswordResetDto": {
        "type": "object",
        "properties": {}
      },
...
      "BachelorDto": {
        "type": "object",
        "properties": {}
      },
      "MoveToTemporarySessionDto": {
        "type": "object",
        "properties": {}
      },
      "TransferLateStudentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}